+++ js/src/index.js
@@ -1,3 +1,5 @@
+import fetch from 'node-fetch';
+import { HttpsProxyAgent } from 'https-proxy-agent';
 /**
  * Sunbird dcTrack API client version %VERSION% in JavaScript
  */
@@ -6,14 +8,16 @@ export class Client {
     #username;
     #password;
     #apiToken;
+    #proxyAgent;
     /**
      * Provide either a username and password, or an API token to access the dcTrack database with JavaScript.
      */
-    constructor(base_url = '', credentials = { username: '', password: '', apiToken: '' }) {
+    constructor(base_url = '', credentials = { username: '', password: '', apiToken: '' }, proxy = '') {
         this.#base_url = base_url;
         this.#username = credentials.username;
         this.#password = credentials.password;
         this.#apiToken = credentials.apiToken;
+        this.#proxyAgent = new HttpsProxyAgent(proxy);
     }
 
     /**
@@ -21,7 +25,7 @@ export class Client {
      */
     async generateToken() {
         if (this.#username && this.#password && !this.#apiToken) {
-            return (await fetch(this.#base_url + '/api/v2/authentication/login', { method: 'POST', headers: [['Authorization', 'Basic ' + btoa(this.#username + ':' + this.#password)]] })).headers.get('Authorization').split(' ')[1];
+            return (await fetch(this.#base_url + '/api/v2/authentication/login', { method: 'POST', agent: this.#proxyAgent, headers: [['Authorization', 'Basic ' + btoa(this.#username + ':' + this.#password)]] })).headers.get('Authorization').split(' ')[1];
         } else {
             throw new Error('Username/password undefined or token predefined.');
         }
@@ -34,6 +38,6 @@ export class Client {
         if (!this.#apiToken) {
             this.#apiToken = await this.generateToken();
         }
-        return (await fetch(this.#base_url + '/' + endpoint, { method: method, headers: [['Authorization', 'Bearer ' + this.#apiToken], ['Content-Type', 'application/json']], body: JSON.stringify(body) })).json();
+        return (await fetch(this.#base_url + '/' + endpoint, { method: method, agent: this.#proxyAgent, headers: [['Authorization', 'Bearer ' + this.#apiToken], ['Content-Type', 'application/json']], body: JSON.stringify(body) })).json();
     }
 }